add data

+++ app/src/main/java/com/heinika/pureweather/json/entity/JsonWeatherEntity.kt
@@ -49,6 +49,6 @@ data class Weather(
 )
 
 data class Wind(
-    val speed: Double,
-    val deg: Double
+    val speed: Float,
+    val deg: Int
 )
